Replace `mem::uninitialized()` with `mem::zeroed()` (#83)
It *probably* doesn't matter.  But really it can only help.  It's theoretically a performance malus but it's so tiny to be irrelevant; I have yet to need a program to find out how many CPU's it has more than once.

Safety: all the uses of `zeroed()` are for C structs that contain integers or `c_char` only, so it's always valid.